LayoutTests/imported/w3c/web-platform-tests/intersection-observer/svg-rect-target-expected.txt


PASS IntersectionObserver observing an SVG <rect> element
FAIL First rAF. assert_approx_equals: entries[0].boundingClientRect.left expected 108 +/- 0 but got 8
FAIL document.scrollingElement.scrollTop = 300 assert_approx_equals: entries[1].boundingClientRect.left expected 108 +/- 0 but got 8
FAIL document.scrollingElement.scrollTop = 100 assert_approx_equals: entries[2].boundingClientRect.left expected 108 +/- 0 but got 8
